private misskey

+++ packages/backend/src/server/api/endpoints/admin/emoji/delete-bulk.ts
@@ -31,6 +31,7 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 		private customEmojiService: CustomEmojiService,
 	) {
 		super(meta, paramDef, async (ps, me) => {
+			if (!me) return null;
 			await this.customEmojiService.deleteBulk(ps.ids, me);
 		});
 	}
