+++ packages/backend/src/server/api/endpoints/admin/emoji/delete.ts
@@ -37,6 +37,7 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 		private customEmojiService: CustomEmojiService,
 	) {
 		super(meta, paramDef, async (ps, me) => {
+			if (!me) return null;
 			await this.customEmojiService.delete(ps.id, me);
 		});
 	}
