+++ packages/backend/src/server/api/endpoints/i/unpin.ts
@@ -46,6 +46,7 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 		private notePiningService: NotePiningService,
 	) {
 		super(meta, paramDef, async (ps, me) => {
+			if (!me) return null;
 			await this.notePiningService.removePinned(me, ps.noteId).catch(err => {
 				if (err.id === 'b302d4cf-c050-400a-bbb3-be208681f40c') throw new ApiError(meta.errors.noSuchNote);
 				throw err;
